requirements.txt
Flask==2.3.3
requests==2.32.3
pandas==2.1.4
numpy==1.26.4
networkx==3.2.1
scikit-learn==1.3.2
python-dotenv==1.0.1
uvicorn==0.23.2
pytest==7.4.4
pytest-cov==4.1.0
